Add docs for u, v, decider on graph flow link method
The docstring is now quite important to know what the
params are (they are no longer as obvious due to the
decider addition) so add a useful docstring to that
method.

This also fixes how the targeted flow add was not passing
allow **kwargs as it should have been and removes the
duplicated docstrings that are in the child class (the
parent class docstrings are just fine).

Change-Id: Idacb7ee9f652fab4bdc762c1b49d7523e46e9a7b