Modify listeners to handle the results now possible from revert()
Now that the REVERT and REVERT_FAILURE states can produce results
or failure objects we need to take that into account in listeners
that were not expecting those states to produce anything; this change
adjusts the built-in listeners so that they now handle these states
and the results they produce.

Also removes some no longer needed py2.6 code used in the logging
listener, as that is not needed anymore since we dropped py2.6
support.

Change-Id: I0d0a9759648b2a2f27a97c68e19c7cdb6375a4f2