Use 'iter_utils.count' to determine how many unfinished nodes left
On state machine 'game_over' reaction callback we currently use
a routine to determine how many nodes are unfinished and then returning
the final state (SUSPENDED) if any are; since we just added the iterator
utils code we can just use the utility function 'count' provided from that
instead.

This explicit usage should make it more clear what this small reaction
callback is doing, and why it returns SUSPENDED.

This also switches it so that in the 'game_over' reaction node deciders
aren't ran (as there is no point in running them in this state).

Change-Id: Ief18598a537d40e5a9b7c50133f158083457755d