Update from global requirements
stable/juno was just created for taskflow from the 0.6.1 tag so that we
can cap networkx (and other requirements) for taskflow in Juno.  This is
needed to unblock the gate-cinder-python26 job in Juno which uses
taskflow and with uncapped networkx, pulls in 1.10 which dropped py26
support.

As the first change to stable/juno for taskflow, this also updates
the .gitreview file.

Depends-On: I29aa97b542d82aabe770ad4431125ef5c67f6104

Closes-Bug: #1484267

Change-Id: Ie5fadc3f26b6897440023b2fdb47a07766061cbb