Use batch 'get_atoms_states' where we can
Instead of doing many single calls to 'get_atom_state' when
we are iterating over many atoms instead prefer to use the
bulk call via 'get_atoms_states' that can gather a large amount
of needed states in a single call, optimizing more of the reading
that taskflow does with regards to states.

This also removes the proxy method 'get_state' from the engine
internal code, since it is no longer needed after this change.

Change-Id: I90eb43e754a7e5efb657468361d67dbe69d31844