Add rundimentary and/or non-optimized job priorities
Allow for providing job priority on job posting and order
iteration by priority (this does not store in any special way
different priorities under different keys/paths, but it could be
done that way in the future).

This is backwards compat in that all old jobs posted without a
priority assume a normal/default priority.

Change-Id: Ie670023d624f4442a16a1278a3aee0d88b102842