Add the ability to skip resolving from activating
It can be useful for debugging a failure to be able
to stop when a failure is detected so that the failure
can be examined in more detail.

This adds the ability to have a 'never_resolve' engine
option which will skip such resolution attempts for
these types of scenarios.

Change-Id: If2386cba970e14c7168a6ff2d4fd4786bcd8e989