import zuul job settings from project-config
This is a mechanically generated patch to complete step 1 of moving
the zuul job settings out of project-config and into each project
repository.

Because there will be a separate patch on each branch, the branch
specifiers for branch-specific jobs have been removed.

See the python3-first goal document for details:
https://governance.openstack.org/tc/goals/stein/python3-first.html

Update the command in the docs environment to not treat warnings as
errors because there are warnings in the documentation build and fixing
them on this old branch will take more work than seems worth investing.

Change-Id: I9e0975919dad547017586378513971f43a8644ad
Story: #2003250