Uncap jsonschema
We have jsonschema capped at a fairly old version. Other than some
specific releases, it looks like keeping it below 3.0 was added in
I943fd68b9fab3bce1764305a5058df5339470757 without really any explanation
why.

In order to update to a 3.x release we need to:
1. Remove the cap from global-requirements.txt (see Depends-On), leaving
   upper-constraints.txt at a 2.x release
2. Remove the cap from all consumers (this change)
3. Release a new version of consumers that are published to pypi
4. Update upper-constraints.txt with those new releases
5. Update jsonschema in upper-constraints.txt to a 3.X release
   (See: https://review.openstack.org/649789)
6. Test consumers with the change from 5.
7. [Optional] fix issues in consumers that arise from 6.
8. Merge the change from 5.

Also includes squashed commit:
Raise lower constraint of psycopg2

Versions prior to 3.7 did not support postgresql version numbering
starting with postgresql 10, causing setup failures. This raises the
lower-constraint to the first version that contained the fix [0] to
support that.

With the change in global requirements for jsonschema, requirements
check fails on raising psycopg2, but we need that raised in order for
the jsonschema version to be raised.

[0] https://github.com/psycopg/psycopg2/pull/504

Also required updating networkx constraints to match current g-r.

Change-Id: Ifd0f0586f945d1da4dd597ae7b4514117cd49f57
Co-Authored-by: Sean McGinnis <sean.mcginnis@gmail.com>
Depends-On: https://review.openstack.org/649669