Raise lower constraint of psycopg2
Versions prior to 3.7 did not support postgresql version numbering
starting with postgresql 10, causing setup failures. This raises the
lower-constraint to the first version that contained the fix [0] to
support that.

With the change in global requirements for jsonschema, requirements
check fails on raising psycopg2, but we need that raised in order for
the jsonschema version to be raised.

[0] https://github.com/psycopg/psycopg2/pull/504

NOTE: Modified from merge to master branch due to the need to squash
both this change with a change to uncap jsonschema to get tests to pass.
This is the extraction of just the psycopg2 change that would have been
done alone in I779ba615c5905eca23ab5074a08483dccc60db10.

(cherry picked from commit 9ea4dae964e590a56d05e0f563c27732a20c2890)

Change-Id: I5000833b802a631e6bb1dbe234ea7d6881e769bf