Update Sphinx requirement
Sphinx 2.0 no longer works on python 2.7, so we need to start capping it
there. We remove Sphinx and doc8 from the 'test-requirements.txt' file
and '[extra] test' section in 'setup.cfg' since none of our tests
actually use this. Finally, we remove a number of dependencies from
'doc/requirements.txt' since these were already specified in
'test-requirements.txt' and we install these as part of the tox target.

Change-Id: I3981f47f4278f726e3c078723f90d4b9ba1699a6
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>