Fix test_while_is_not with python 3.11
The test_while_is_not function relied on strings and literals to test
the 'while_is_not' function. But the while_is_not function uses the 'is'
operator to test the equivalency of 2 objects.
This triggers a bug with python 3.11 where using 'is' with literals is
not advised (it is not recommended since python 3.8 [0]).
The test now uses objects from a specific class to evaluate the
while_is_not function.

[0] https://docs.python.org/3/whatsnew/3.8.html#changes-in-python-behavior

Change-Id: I38a0135aaf73e25aa20a11c0685d5c2a7b587a07