Use consistent credential for Redis and Redis Sentinel
We generally expects that users enabled authentication both in Redis
and Redis Sentinel at the same time, and it'd be more reasonable to
use the same credential by default. This behavior is more consistent
with the other libraries used in tooz or dogpile.cache, which are used
in multiple OpenStack services.

Change-Id: Ie8c96415f72bbd5f11f908bb9f96242d57b2bd4e